Implement function percentile in tool dudect
Because of the unstable constent time test result when run the "make test"
command, I compared the dudect tool implementation and the source file in
paper <Dude,is my code constant time?>. Find out that the function
percentile which is used to crop the outliers was missing. Therefore,
I insert the missing function and change the data structure "t_context_t"
to fit the implementation of percentile.