Fix the segmentation fault in descend_ascend
During the simplifications of functions q_ascend and q_descend,
the additional function descend_ascend occur segmentation fault.
The reseasion is that the stop condition of the loop which travel
through queue can not be the "node" in list function
"list_for_each_safe" because the while loop inside for loop may
delete multiple node. Therefore, the new for loop set the stop
condition on "safe".